refactor drawing to support layers
Drawing elements that appear atop or below layers need to be
drawn both in the proper order and with the correct canvas to
respect clipping and the matrix.

Drawing the find results, text selection, or the cursor ring,
interleaves with any layers that may be drawn. The main picture
is treated as owned by a LayerAndroid so each component can
decide when to draw.

This change leave the main picture in WebViewCore.cpp, and
draws everything else in WebView.cpp -- in the future, additional
refactoring can put all drawing in one place.

The logic of what to draw is still in WebView.java, but the
actual drawing calls are now triggered inside the layer code.

Android.mk
- Add rule to trigger building without layers from buildspec.mk.

LayerAndroid.*
- Replace FindOnPage reference with abstract DrawExtra class to
  draw adornments in the layers' canvas context.
- Add a LayerAndroid constructor to create a dummy layer with a
  SkPicture* and a uniqueId==-1 so that extras can detect when
  they are drawn by the main picture.

android_graphics.*
- Move cursor ring drawing out of WebView.cpp to here.
- Separate cursor ring setup from actual drawing.
- Get the cursor ring metrics in local coordinates.

ChromeClientAndroid.cpp
- Fix compiler warnings.

WebViewCore.*
- Move updateCursorBounds from WebView.cpp. This permits it to
  be called from CursorRing::setup.

CachedFrame.*
CachedNode.*
CachedLayer.*
- Add local bounds getters.

CachedRoot.h
- Move class FindCanvas to the android namespace.

DrawExtra.h
- Add an abstract class called by LayerAndroid to optionally
  draw extra elements in its canvas context.

FindCanvas.*
SelectText.*
- Refactor drawing to draw in layers context.

WebView.cpp
- Move drawing from WebView.java.
- Remove selection code to SelectText.cpp.
- Use inverseScale to simplify viewPort metrics.
- Simplify layer root so java doesn't need to know about it.

Requires companion change in frameworks/base

http://b/2457316
http://b/2454127
http://b/2454149

+++ WebKit/android/WebCoreSupport/ChromeClientAndroid.cpp
@@ -374,7 +374,7 @@ void ChromeClientAndroid::exceededDatabaseQuota(Frame* frame, const String& name
     // If new quota is unavailable, we may be able to resolve the situation by shrinking the quota of an origin that asked for a lot but is only using a little.
     // If we find such a site, shrink it's quota and ask Java to try again.
 
-    if (m_newQuota == currentQuota && !m_triedToReclaimDBQuota) {
+    if ((unsigned long long) m_newQuota == currentQuota && !m_triedToReclaimDBQuota) {
         m_triedToReclaimDBQuota = true; // we should only try this once per quota overflow.
         unsigned long long reclaimedQuotaBytes = tryToReclaimDatabaseQuota(origin);
 
@@ -397,7 +397,7 @@ static unsigned long long tryToReclaimDatabaseQuota(SecurityOrigin* originNeedin
     Vector<RefPtr<SecurityOrigin> > origins;
     tracker.origins(origins);
     unsigned long long reclaimedQuotaBytes = 0;
-    for (int i = 0; i < origins.size(); i++) {
+    for (unsigned i = 0; i < origins.size(); i++) {
         SecurityOrigin* originToReclaimFrom = origins[i].get();
 
         // Don't try to reclaim from the origin that has exceeded its quota.
